chore: Add _fluttefire_internals and extract firebase_core/src/internals to the new package.
