+++ packages/_flutterfire_internals/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 1.0.0
+
+Initial release of \_fluttefire_internals
