+++ packages/_flutterfire_internals/README.md
@@ -1,0 +1,4 @@
+# \_flutterfire_internals
+
+This package is hosting Dart code shared between FlutterFire plugins. It is not
+meant for public consumption.
