+++ packages/_flutterfire_internals/lib/src/interop_shimmer.dart
