+++ packages/_flutterfire_internals/pubspec.yaml
@@ -1,0 +1,25 @@
+name: _flutterfire_internals
+description: A package hosting Dart code shared between FlutterFire plugins.
+homepage: https://firebase.google.com/docs/firestore
+repository: https://github.com/firebase/flutterfire/tree/master/packages/_flutterfire_internals
+version: 1.0.0
+
+environment:
+  sdk: '>=2.16.0 <3.0.0'
+  flutter: '>=1.12.13+hotfix.5'
+
+dependencies:
+  cloud_firestore_platform_interface: ^5.7.5
+  cloud_firestore_web: ^2.8.8
+  collection: ^1.0.0
+  firebase_core: ^1.10.2
+  firebase_core_platform_interface: ^4.5.1
+  flutter:
+    sdk: flutter
+  meta: ^1.3.0
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  mockito: ^5.0.0
+  test: any
