+++ packages/_flutterfire_internals/test/guard_test.dart
@@ -1,5 +1,5 @@
-import 'package:firebase_core/src/internals.dart';
-import 'package:firebase_core/src/interop_shimmer.dart';
+import 'package:_flutterfire_internals/_flutterfire_internals.dart';
+import 'package:_flutterfire_internals/src/interop_shimmer.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
