+++ packages/cloud_firestore/cloud_firestore_web/pubspec.yaml
@@ -10,6 +10,7 @@ environment:
   flutter: ">=1.12.13+hotfix.4"
 
 dependencies:
+  _flutterfire_internals: ^1.0.0
   cloud_firestore_platform_interface: ^5.7.5
   collection: ^1.0.0
   firebase_core: ^1.10.0
