+++ packages/firebase_analytics/firebase_analytics_web/lib/utils/exception.dart
@@ -1,8 +1,6 @@
-export 'package:firebase_core/src/internals.dart' hide guardWebExceptions;
-
 import 'package:firebase_core/firebase_core.dart';
-// ignore: implementation_imports
-import 'package:firebase_core/src/internals.dart' as internals;
+import 'package:_flutterfire_internals/_flutterfire_internals.dart'
+    as internals;
 
 /// Will return a [FirebaseException] from a thrown web error.
 /// Any other errors will be propagated as normal.
