+++ packages/firebase_analytics/firebase_analytics_web/pubspec.yaml
@@ -9,6 +9,7 @@ environment:
   flutter: '>=1.12.13+hotfix.4'
 
 dependencies:
+  _flutterfire_internals: ^1.0.0
   firebase_analytics_platform_interface: ^3.3.5
   firebase_core: ^1.10.0
   firebase_core_web: ^1.7.2
