+++ packages/firebase_app_check/firebase_app_check_web/pubspec.yaml
@@ -8,6 +8,7 @@ environment:
   flutter: ">=1.20.0"
 
 dependencies:
+  _flutterfire_internals: ^1.0.0
   firebase_app_check_platform_interface: ^0.0.4+17
   firebase_core: ^1.10.0
   firebase_core_web: ^1.7.2
