+++ packages/firebase_app_installations/firebase_app_installations_web/lib/src/guard.dart
@@ -4,10 +4,8 @@
 
 import 'package:firebase_core/firebase_core.dart';
 
-// ignore: implementation_imports
-import 'package:firebase_core/src/internals.dart' as internals;
-
-export 'package:firebase_core/src/internals.dart' hide guardWebExceptions;
+import 'package:_flutterfire_internals/_flutterfire_internals.dart'
+    as internals;
 
 /// Will return a [FirebaseException] from a thrown web error.
 /// Any other errors will be propagated as normal.
