+++ packages/firebase_app_installations/firebase_app_installations_web/pubspec.yaml
@@ -9,6 +9,7 @@ environment:
   flutter: '>=1.20.0'
 
 dependencies:
+  _flutterfire_internals: ^1.0.0
   firebase_app_installations_platform_interface: ^0.1.1+17
   firebase_core: ^1.10.0
   firebase_core_web: ^1.7.2
