+++ packages/firebase_messaging/firebase_messaging_web/lib/src/internals.dart
@@ -1,9 +1,6 @@
-// ignore_for_file: require_trailing_commas
-export 'package:firebase_core/src/internals.dart' hide guardWebExceptions;
-
 import 'package:firebase_core/firebase_core.dart';
-// ignore: implementation_imports
-import 'package:firebase_core/src/internals.dart' as internals;
+import 'package:_flutterfire_internals/_flutterfire_internals.dart'
+    as internals;
 
 /// Will return a [FirebaseException] from a thrown web error.
 /// Any other errors will be propagated as normal.
