+++ packages/firebase_messaging/firebase_messaging_web/pubspec.yaml
@@ -9,6 +9,7 @@ environment:
   flutter: ">=1.12.13+hotfix.4"
 
 dependencies:
+  _flutterfire_internals: ^1.0.0
   firebase_core: ^1.10.0
   firebase_core_web: ^1.7.2
   firebase_messaging_platform_interface: ^4.1.5
