+++ packages/firebase_performance/firebase_performance_web/lib/src/internals.dart
@@ -2,11 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'package:firebase_core/src/internals.dart' hide guardWebExceptions;
-
 import 'package:firebase_core/firebase_core.dart';
-// ignore: implementation_imports
-import 'package:firebase_core/src/internals.dart' as internals;
+import 'package:_flutterfire_internals/_flutterfire_internals.dart'
+    as internals;
 
 /// Will return a [FirebaseException] from a thrown web error.
 /// Any other errors will be propagated as normal.
