+++ packages/firebase_performance/firebase_performance_web/pubspec.yaml
@@ -8,6 +8,7 @@ environment:
   flutter: ">=1.20.0"
 
 dependencies:
+  _flutterfire_internals: ^1.0.0
   firebase: ^9.0.1
   firebase_core: ^1.10.0
   firebase_core_web: ^1.7.2
