+++ packages/firebase_storage/firebase_storage_web/lib/src/utils/errors.dart
@@ -3,9 +3,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:_flutterfire_internals/_flutterfire_internals.dart'
+    as internals;
 import 'package:firebase_core/firebase_core.dart';
-// ignore: implementation_imports
-import 'package:firebase_core/src/internals.dart' as internals;
 
 Map<String, String?> _errorCodeToMessage = {
   'unauthorized': 'User is not authorized to perform the desired action.',
