+++ packages/firebase_storage/firebase_storage_web/pubspec.yaml
@@ -9,6 +9,7 @@ environment:
   flutter: ">=1.20.0"
 
 dependencies:
+  _flutterfire_internals: ^1.0.0
   async: ^2.5.0
   firebase_core: ^1.10.0
   firebase_core_web: ^1.7.2
